+++ sync-for-payroll/CodatSyncPayroll/Models/Operations/ListTrackingCategoriesResponse.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncPayroll.Models.Operations
+namespace Codat.Sync.Payroll.Models.Operations
 {
-    using CodatSyncPayroll.Models.Shared;
+    using Codat.Sync.Payroll.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -35,11 +35,11 @@ namespace CodatSyncPayroll.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
 
         /// <summary>
         /// Success
         /// </summary>
-        public TrackingCategories? TrackingCategories { get; set; }
+        public Models.Shared.TrackingCategories? TrackingCategories { get; set; }
     }
 }